moved export logics to synchronization
